refactor: FSD 규칙에 어긋나는 디펜던시 관계 정리
