+++ src/entity/diet/api/queries.ts
@@ -1,32 +1,9 @@
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 
 import { authApi } from '@/entity/auth';
-import { Comment } from '@/feature/diet'; //TODO: Feature ->entity로 이동
 import { BaseError, BaseResponse, Pageable } from '@/shared/api';
 
 import { HomeDietData } from '../model/types';
-
-interface DietCommentRequest {
-  dietId: number;
-  size: number;
-}
-
-export const useDietCommentListQuery = ({ dietId, size }: DietCommentRequest) => {
-  return useInfiniteQuery<Comment, BaseError>({
-    queryKey: ['dietCommentList', dietId],
-    queryFn: async ({ pageParam }) => {
-      const result = await authApi.get<BaseResponse<Comment>>(
-        `/api/diets/v1/${dietId}/comments?page=${pageParam as number}&size=${size}`
-      );
-      return result.data.data;
-    },
-    initialPageParam: 0,
-
-    getNextPageParam: (lastPage: Comment, allPages: Comment[]) => {
-      return lastPage.content ? allPages.length : undefined;
-    },
-  });
-};
 
 interface DietRequest {
   searchDate: string;
